+++ packages/ckeditor5-list/theme/listproperties.css
@@ -4,11 +4,58 @@
  */
 
 .ck.ck-list-properties {
+	/* When there are no list styles and there is no collapsible. */
+	&.ck-list-properties_without-styles {
+		padding: var(--ck-spacing-large);
+
+		& > * {
+			min-width: 14em;
+
+			& + * {
+				margin-top: var(--ck-spacing-standard);
+			}
+		}
+	}
+
 	/*
 	 * When the numbered list property fields (start at, reversed) should be displayed,
 	 * more horizontal space is needed. Reconfigure the style grid to create that space.
 	 */
-	&.ck-list-properties_with-numbered-properties > .ck-list-styles-list {
-		grid-template-columns: repeat( 4, auto );
+	&.ck-list-properties_with-numbered-properties {
+		& > .ck-list-styles-list {
+			grid-template-columns: repeat( 4, auto );
+		}
+
+		/* When list styles are rendered and property fields are in a collapsible. */
+		& > .ck-collapsible {
+			border-top: 1px solid var(--ck-color-base-border);
+
+			& > .ck-collapsible__children {
+				& > * {
+					width: 100%;
+
+					& + * {
+						margin-top: var(--ck-spacing-standard);
+					}
+				}
+			}
+		}
+	}
+
+	& .ck.ck-numbered-list-properties-start-index .ck-input {
+		min-width: auto;
+		width: 100%;
+	}
+
+	& .ck.ck-numbered-list-properties-reversed-order {
+		background: transparent;
+		padding-left: 0;
+		padding-right: 0;
+		padding-bottom: 0;
+
+		&:active {
+			box-shadow: none;
+			border-color: transparent;
+		}
 	}
 }
