--- packages/ckeditor5-list/theme/numberedlistproperties.css
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2003-2021, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-
-.ck.ck-numbered-list-properties {
-	& .ck.ck-numbered-list-properties-start-index .ck-input {
-		min-width: auto;
-		width: 100%;
-	}
-
-	& .ck.ck-numbered-list-properties-reversed-order {
-		background: transparent;
-		padding-left: 0;
-		padding-right: 0;
-		padding-bottom: 0;
-
-		&:active {
-			box-shadow: none;
-			border-color: transparent;
-		}
-	}
-
-	/* When there are no list styles and there is no collapsible. */
-	&:only-child {
-		padding: var(--ck-spacing-large);
-
-		& > * {
-			min-width: 14em;
-
-			& + * {
-				margin-top: var(--ck-spacing-standard);
-			}
-		}
-	}
-
-	/* When list styles are rendered and property fields are in a collapsible. */
-	& > .ck-collapsible {
-		border-top: 1px solid var(--ck-color-base-border);
-
-		& > .ck-collapsible__children {
-			& > * {
-				width: 100%;
-
-				& + * {
-					margin-top: var(--ck-spacing-standard);
-				}
-			}
-		}
-	}
-}
